Foundations: Control Flow #time 2m

+++ control-flow/solution/operators_even_odd.py
@@ -1,7 +1,7 @@
 number = input('Podaj liczbę: ')
 
 
-if float(number) % 2 == 0:
+if int(number) % 2 == 0:
     print('Even')
 else:
     print('Odd')
